extract resource methods
